@@ -2,7 +2,7 @@ package com.example.eziketobenna.bakingapp.data.network;
 
 import android.app.IntentService;
 import android.content.Intent;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.Log;
 
 import com.example.eziketobenna.bakingapp.BakingApplication;
