@@ -1,8 +1,8 @@
 package com.example.eziketobenna.bakingapp.data.database;
 
-import android.arch.persistence.room.Database;
-import android.arch.persistence.room.RoomDatabase;
-import android.arch.persistence.room.TypeConverters;
+import androidx.room.Database;
+import androidx.room.RoomDatabase;
+import androidx.room.TypeConverters;
 
 import com.example.eziketobenna.bakingapp.data.model.Recipe;
 
